Allow base_container.sh to push to registry
Before running this script, one should spin up the registry as the
vagrant user. Running this during bootstrap seems to mess up the
user namespace podman uses in order to run root commands.

The base_container.sh script also creates a config to go along
with the sbom. This may or may not be useful later.

Signed-off-by: Nisha K <nishak@vmware.com>